+++ include/continental/hydrotools/service/StreamAccumLength.h
@@ -31,19 +31,19 @@ class StreamAccumLength
 {
 
 private:
-    std::shared_ptr<datamanagement::Raster<short>> m_strDef;
-    std::shared_ptr<datamanagement::Raster<short>> m_flowDirection;
+    std::shared_ptr<datamanagement::Raster<float>> m_strDef;
+    std::shared_ptr<datamanagement::Raster<float>> m_flowDirection;
     std::shared_ptr<datamanagement::Raster<float>> m_accumLength;
-    std::shared_ptr<datamanagement::Raster<short>> m_strSeg;
+    std::shared_ptr<datamanagement::Raster<float>> m_strSeg;
 
 public:
     StreamAccumLength();
 
     //Lê os dados do FlowDirection
-    void setFlowDirection(std::shared_ptr<datamanagement::Raster<short>> flowDirection);
+    void setFlowDirection(std::shared_ptr<datamanagement::Raster<float>> flowDirection);
 
     //Lê os dados do FlowDirection
-    void setStreamDefinition(std::shared_ptr<datamanagement::Raster<short>> streamDefinition);
+    void setStreamDefinition(std::shared_ptr<datamanagement::Raster<float>> streamDefinition);
 
     /// <summary>
     /// Separa os trechos de rio em todas as confluências
@@ -58,7 +58,7 @@ public:
 
 private:
     //Atualiza o atributo do stream definition
-    void updateReaches(int row, int column, int rowLast, int columnLast, short attribute, int &segmentNumberIncrease, float nReaches, float reachLength);
+    void updateReaches(int row, int column, int rowLast, int columnLast, float attribute, int &segmentNumberIncrease, float nReaches, float reachLength);
 
     bool headwaterFound(int yc, int xc, bool searchAccumLength);
