+++ src/service/StreamAccumLength.cpp
@@ -15,15 +15,15 @@ namespace service
 
 StreamAccumLength::StreamAccumLength()
 {
-    m_flowDirection = make_shared<Raster<short>>();
+    m_flowDirection = make_shared<Raster<float>>();
 }
 
-void StreamAccumLength::setFlowDirection(shared_ptr<Raster<short>> flowDirection)
+void StreamAccumLength::setFlowDirection(shared_ptr<Raster<float>> flowDirection)
 {
     m_flowDirection = flowDirection;
 }
 
-void StreamAccumLength::setStreamDefinition(shared_ptr<Raster<short>> streamDefinition)
+void StreamAccumLength::setStreamDefinition(shared_ptr<Raster<float>> streamDefinition)
 {
     m_strDef = streamDefinition;
     m_accumLength = make_shared<Raster<float>>(m_strDef->getRows(), m_strDef->getCols(), m_strDef->getXOrigin(), m_strDef->getYOrigin(), m_strDef->getCellSize(), m_strDef->getNoDataValue());
@@ -125,7 +125,7 @@ void StreamAccumLength::segmentStreamsByLength(float MaxLength)
 {
     // A matriz do stream definition é liberada para otimizar a memória
     m_strDef = nullptr;
-    m_strSeg = make_shared<Raster<short>>(m_flowDirection->getRows(), m_flowDirection->getCols(), m_flowDirection->getXOrigin(), m_flowDirection->getYOrigin(), m_flowDirection->getCellSize(), m_flowDirection->getNoDataValue());
+    m_strSeg = make_shared<Raster<float>>(m_flowDirection->getRows(), m_flowDirection->getCols(), m_flowDirection->getXOrigin(), m_flowDirection->getYOrigin(), m_flowDirection->getCellSize(), m_flowDirection->getNoDataValue());
 
     float TotalLength = 0;
     int SegmentNumber = 1;
@@ -221,7 +221,7 @@ void StreamAccumLength::segmentStreamsByLength(float MaxLength)
     }
 }
 
-void StreamAccumLength::updateReaches(int linha, int coluna, int LinhaFinal, int ColunaFinal, short Atributo, int &SegmentNumberIncrease, float nReaches, float ReachLength)
+void StreamAccumLength::updateReaches(int linha, int coluna, int LinhaFinal, int ColunaFinal, float Atributo, int &SegmentNumberIncrease, float nReaches, float ReachLength)
 {
     //0 = célula ainda não identificada
